+++ src/Index.jsx
@@ -2,13 +2,10 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import App from './App';
 
-import { createStore } from 'redux';
 import { Provider } from 'react-redux';
-import reducer from './redux/reducer';
+import store from './redux/reducer';
 
 import './Index.less';
-
-const store = createStore(reducer);
 
 ReactDOM.render(
   <Provider store={store}>