+++ src/redux/uiState/index.js
@@ -1,0 +1,6 @@
+import slice from 'Redux/uiState/slice';
+import * as uiStateSelectors from 'Redux/uiState/selectors';
+
+const { actions, reducer } = slice;
+
+export { actions, reducer, uiStateSelectors };